+++ docker-api/docker-api/main.go
@@ -82,7 +82,6 @@ type config struct {
 	LogsTimestamps          bool
 	LogsDetails             bool
 	Monolithic              bool
-	VolumeMounts            []volumeMount
 }
 
 type dockerConfig struct {
@@ -93,11 +92,6 @@ type registryCredential struct {
 	Username string
 	Password string
 	Address  string
-}
-
-type volumeMount struct {
-	Source string
-	Target string
 }
 
 const defaultMaxRecvMsgSize = 5 * 1024 * 1024 * 1024